+++ cmd/server/cli/generate.go
@@ -2,6 +2,7 @@ package cli
 
 import (
 	"fmt"
+
 	"github.com/HewlettPackard/galadriel/cmd/server/util"
 	"github.com/google/uuid"
 	"github.com/spf13/cobra"
@@ -14,7 +15,7 @@ var generateCmd = &cobra.Command{
 var tokenCmd = &cobra.Command{
 	Use:   "token <memberID>",
 	Args:  cobra.ExactArgs(1),
-	Short: "Generates a join token for provided member",
+	Short: "Generates an access token for provided memberID",
 	RunE: func(cmd *cobra.Command, args []string) error {
 		c := util.NewServerClient(defaultSocketPath)
 		memberID, err := uuid.Parse(args[0])
