+++ pkg/server/datastore/datastore.go
@@ -2,15 +2,17 @@ package datastore
 
 import (
 	"context"
+	"errors"
 	"fmt"
-	"github.com/HewlettPackard/galadriel/pkg/common"
-	"github.com/google/uuid"
 	"sync"
 	"time"
+
+	"github.com/HewlettPackard/galadriel/pkg/common"
+	"github.com/google/uuid"
 )
 
 type DataStore interface {
-	CreateMember(ctx context.Context, member *common.Member) (*Member, error)
+	CreateMember(ctx context.Context, m *common.Member) (*Member, error)
 	CreateRelationship(ctx context.Context, r *common.Relationship) (*Relationship, error)
 	CreateAccessToken(ctx context.Context, t *common.AccessToken, memberID uuid.UUID) (*AccessToken, error)
 }
@@ -38,15 +40,17 @@ type Relationship struct {
 // TODO: use until an actual DataStore implementation is added.
 
 type MemStore struct {
-	member       []Member
-	relationship []Relationship
+	member       map[uuid.UUID]*Member
+	relationship map[uuid.UUID]*Relationship
 
 	mu sync.RWMutex
 }
 
 func NewMemStore() DataStore {
 	return &MemStore{
-		mu: sync.RWMutex{},
+		member:       make(map[uuid.UUID]*Member),
+		relationship: make(map[uuid.UUID]*Relationship),
+		mu:           sync.RWMutex{},
 	}
 }
 
@@ -54,55 +58,47 @@ func (s *MemStore) CreateMember(_ context.Context, member *common.Member) (*Memb
 	s.mu.Lock()
 	defer s.mu.Unlock()
 
-	tokens := make([]AccessToken, 0)
-	for _, t := range member.Tokens {
-		tokens = append(tokens, AccessToken{Token: t.Token})
-	}
-	id, err := uuid.NewUUID()
-	if err != nil {
-		return nil, err
-	}
 	m := Member{
-		ID:          id,
+		ID:          uuid.New(),
 		Name:        member.Name,
 		TrustDomain: member.TrustDomain,
-		Tokens:      tokens,
 	}
-	s.member = append(s.member, m)
+
+	for _, t := range member.Tokens {
+		m.Tokens = append(m.Tokens, AccessToken{Token: t.Token, Expiry: t.Expiry})
+	}
+	s.member[m.ID] = &m
+
 	fmt.Println("Members:", s.member)
-	return &m, nil
+	return s.member[m.ID], nil
 }
 
 func (s *MemStore) CreateRelationship(_ context.Context, rel *common.Relationship) (*Relationship, error) {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 
-	s.relationship = append(s.relationship, Relationship{
+	r := Relationship{
+		ID:      uuid.New(),
 		MemberA: rel.MemberB,
 		MemberB: rel.MemberB,
-	})
-	return &s.relationship[len(s.relationship)-1], nil
+	}
+	s.relationship[r.ID] = &r
+
+	fmt.Println("Relationships:", s.relationship)
+	return s.relationship[r.ID], nil
 }
 
 func (s *MemStore) CreateAccessToken(_ context.Context, t *common.AccessToken, memberID uuid.UUID) (*AccessToken, error) {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 
-	i, err := s.getMemberIndex(memberID)
-	if err != nil {
-		return nil, err
+	m, ok := s.member[memberID]
+	if !ok {
+		return nil, errors.New("member not found")
 	}
 
-	s.member[i].Tokens = append(s.member[i].Tokens, AccessToken{Token: t.Token})
-	// TODO: what to return here?
-	return &s.member[i].Tokens[len(s.member[i].Tokens)-1], nil
-}
+	m.Tokens = append(m.Tokens, AccessToken{Token: t.Token, Expiry: t.Expiry})
+	s.member[memberID] = m
 
-func (s *MemStore) getMemberIndex(id uuid.UUID) (int, error) {
-	for i, m := range s.member {
-		if m.ID == id {
-			return i, nil
-		}
-	}
-	return 0, fmt.Errorf("member not found")
+	return &m.Tokens[len(m.Tokens)-1], nil
 }
