+++ pkg/server/endpoints/management.go
@@ -3,116 +3,13 @@ package endpoints
 import (
 	"context"
 	"encoding/json"
-	"fmt"
-	"github.com/HewlettPackard/galadriel/pkg/common"
-	"github.com/HewlettPackard/galadriel/pkg/common/util"
-	"github.com/HewlettPackard/galadriel/pkg/server/datastore"
-	"github.com/labstack/echo/v4"
-	"github.com/sirupsen/logrus"
 	"io"
-	"net"
 	"net/http"
 	"time"
+
+	"github.com/HewlettPackard/galadriel/pkg/common"
+	"github.com/HewlettPackard/galadriel/pkg/common/util"
 )
-
-// Server manages the UDS and TCP endpoints lifecycle
-type Server interface {
-	// ListenAndServe starts all endpoint servers and blocks until the context
-	// is canceled or any of the endpoints fails to run.
-	ListenAndServe(ctx context.Context) error
-}
-
-type EndpointHandler struct {
-	TCPAddress *net.TCPAddr
-	LocalAddr  net.Addr
-	DataStore  datastore.DataStore
-	Log        logrus.FieldLogger
-}
-
-func New(c Config) (*EndpointHandler, error) {
-	if err := util.PrepareLocalAddr(c.LocalAddress); err != nil {
-		return nil, err
-	}
-	return &EndpointHandler{
-		TCPAddress: c.TCPAddress,
-		LocalAddr:  c.LocalAddress,
-		DataStore:  c.Catalog.GetDataStore(),
-		Log:        c.Log,
-	}, nil
-}
-
-func (e *EndpointHandler) ListenAndServe(ctx context.Context) error {
-	tasks := []func(context.Context) error{
-		e.runTCPServer,
-		e.runUDSServer,
-	}
-
-	err := util.RunTasks(ctx, tasks)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (e *EndpointHandler) runTCPServer(ctx context.Context) error {
-	server := echo.New()
-
-	e.Log.Info("Starting TCP Server")
-	errChan := make(chan error)
-	go func() {
-		errChan <- server.Start(e.TCPAddress.String())
-	}()
-
-	var err error
-	select {
-	case err = <-errChan:
-		e.Log.WithError(err).Error("TCP Server stopped prematurely")
-		return err
-	case <-ctx.Done():
-		e.Log.Info("Stopping TCP Server")
-		server.Close()
-		<-errChan
-		e.Log.Info("TCP Server stopped")
-		return nil
-	}
-}
-
-func (e *EndpointHandler) runUDSServer(ctx context.Context) error {
-	server := &http.Server{}
-
-	l, err := net.Listen(e.LocalAddr.Network(), e.LocalAddr.String())
-	if err != nil {
-		return fmt.Errorf("error listening on uds: %w", err)
-	}
-	defer l.Close()
-
-	e.addHandlers(ctx)
-
-	e.Log.Info("Starting UDS Server")
-	errChan := make(chan error)
-	go func() {
-		errChan <- server.Serve(l)
-	}()
-
-	select {
-	case err = <-errChan:
-		e.Log.WithError(err).Error("Local Server stopped prematurely")
-		return err
-	case <-ctx.Done():
-		e.Log.Info("Stopping UDS Server")
-		server.Close()
-		<-errChan
-		e.Log.Info("UDS Server stopped")
-		return nil
-	}
-}
-
-func (e *EndpointHandler) addHandlers(ctx context.Context) {
-	e.createMemberHandler(ctx)
-	e.createRelationshipHandler(ctx)
-	e.generateTokenHandler(ctx)
-}
 
 func (e *EndpointHandler) createMemberHandler(ctx context.Context) {
 	http.HandleFunc("/createMember", func(w http.ResponseWriter, r *http.Request) {
@@ -195,7 +92,7 @@ func (e *EndpointHandler) createRelationshipHandler(ctx context.Context) {
 }
 
 func (e *EndpointHandler) generateTokenHandler(ctx context.Context) {
-	http.HandleFunc("/token", func(w http.ResponseWriter, r *http.Request) {
+	http.HandleFunc("/createToken", func(w http.ResponseWriter, r *http.Request) {
 		body, err := io.ReadAll(r.Body)
 		if err != nil {
 			e.Log.Errorf("failed parsing body: %v", err)
@@ -211,6 +108,12 @@ func (e *EndpointHandler) generateTokenHandler(ctx context.Context) {
 		}
 
 		token, err := util.GenerateToken()
+		if err != nil {
+			e.Log.Errorf("failed generating token: %v", err)
+			w.WriteHeader(500)
+			return
+		}
+
 		at, err := e.DataStore.CreateAccessToken(
 			ctx, &common.AccessToken{Token: token, Expiry: time.Now()}, m.ID,
 		)
